+++ changelog.md
@@ -26,6 +26,14 @@
 
   local s = t and t.x or 1 -- `t` in `t.x` is `table`
   ```
+* `CHG` infer type by `type(x)`
+  ```lua
+  local x
+
+  if type(x) == 'string' then
+    print(x) -- `x` is `string` here
+  end
+  ```
 * `FIX` with clients that support LSP 3.17 (VSCode), workspace diagnostics are triggered every time when opening a file.
 * `FIX` [#1204](https://github.com/sumneko/lua-language-server/issues/1204)
 * `FIX` [#1208](https://github.com/sumneko/lua-language-server/issues/1208)